# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Documentation API MYAPPS DASHBOARD",
        "description": "Ini Dokumentasi API untuk semua respon yang didapatkan dari API",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:8090",
    "basePath": "/api/",
    "tags": [{
        "name": "API MYAPPS DASHBOARD",
        "description": "Get Data API MYAPPS DASHBOARD"
    }],
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/todo-list": {
            "parameters": [{
                "name": "Authorization",
                "in": "header",
                "required": true,
                "description": "Token Authorization",
                "type": "string"
            }],
            "get": {
                "tags": [
                    "Todo List"
                ],
                "summary": "Get Todo List",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/todo-list-monitoring": {
            "parameters": [{
                "name": "Authorization",
                "in": "header",
                "required": true,
                "description": "Token Authorization",
                "type": "string"
            }],
            "get": {
                "tags": [
                    "Todo List Monitoring"
                ],
                "summary": "Get Todo List Monitoring",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        }
    }
}